msm: camera: Fix memory leak.
CAF never deattaches the dma buffer, creating a memory leak
every time this function is run. This issue becomes increasingly
problematic as it's run a metric ton every time camera is open,
which on Coral is pretty much every time you unlock the phone
and often while you run it too. Fix it by unmaping and detaching
the dma buffer.

Signed-off-by: Demon000 <demonsingur@gmail.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I73e0d27fc0e00effd2507ca3b179c90013b6a319
Signed-off-by: HafizZiq <hafizuddinismail552@gmail.com>